+++ src/testHelpers/createMockUser.ts
@@ -3,7 +3,7 @@
 // Create a mock AuthUser by passing mock input to the
 // real createAuthUser module.
 
-import { createMockFirebaseUserAdminSDK } from 'src/testHelpers/authUserInputs'
+import { createMockFirebaseUserAdminSDK } from 'src/testHelpers/userInputs'
 
 const createAuthUser = jest.requireActual('../createUser').default
